+++ js/gallery.js
@@ -1,0 +1,1 @@
+//Просмотр загруженных изображений
