+++ js/upload.js
@@ -1,0 +1,1 @@
+//Загрузка нового изображения
